The view was implemented to show the direction and speed of the wind

+++ app/src/main/java/com/example/weathercomposeapp/data/model/DailyForecast.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class DailyForecast(
     @SerializedName("Date") val date: String,
     @SerializedName("Day") val day: Day,
-    @SerializedName("EpochDate") val epochDate: Int,
+    @SerializedName("EpochDate") val epochDate: Long,
     @SerializedName("Link") val link: String,
     @SerializedName("MobileLink") val mobileLink: String,
     @SerializedName("Night") val night: Night,
